+++ src/main/java/com/example/seoulcultureport/dto/boardDto/BoardListResponseDto.java
@@ -1,50 +1,38 @@
-package com.example.seoulcultureport.dto;
+package com.example.seoulcultureport.dto.boardDto;
 
 import com.example.seoulcultureport.entity.Board;
-import com.example.seoulcultureport.entity.Comment;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 import java.time.format.DateTimeFormatter;
-import java.util.ArrayList;
-import java.util.List;
 
 @Getter
-public class BoardDetailResponseDto {
+@NoArgsConstructor
+public class BoardListResponseDto {
     private Long id;
     private String title;
     private String image;
-    private String pageUrl;
-    private String classify;
-    private String region;
     private String location;
+    private String region;
     private String startDate;   // 타입 : string
     private String endDate;
-    private String contents;
     private int cmtCount;
-    private String username;
+    private String nickname;
     private String createdat;
 
-    private final List<CommentResponseDto> commentList = new ArrayList<>();
-
-    public BoardDetailResponseDto(Board board) {
+    public BoardListResponseDto(Board board) {
         DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm");
         this.id = board.getId();
         this.title = board.getTitle();
         this.image = board.getImage();
-        this.pageUrl = board.getPageUrl();
-        this.classify = board.getClassify();
-        this.region = board.getRegion();
         this.location = board.getLocation();
+        this.region = board.getRegion();
         this.startDate = board.getStartDate();
         this.endDate = board.getEndDate();
-        this.contents = board.getContents();
         this.cmtCount = board.getCmtCount();
-        this.username = board.getUsername();
+        this.nickname = board.getNickname();
         this.createdat = board.getCreatedAt().format(formatter);
 
-        for (Comment comment : board.getComments()) {
-            commentList.add(new CommentResponseDto(comment));
-        }
     }
 
 }
