+++ src/main/java/com/example/seoulcultureport/dto/boardDto/BoardRequestDto.java
@@ -1,4 +1,4 @@
-package com.example.seoulcultureport.dto;
+package com.example.seoulcultureport.dto.boardDto;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
